+++ CHANGELOG.md
@@ -19,13 +19,43 @@ Security in case of vulnerabilities.
 ### Added
 - Sponsored support grant task for conversion and transfer project.
 
+- Added `Check and confirm the academy and trust financial information` task for transfer project.
+
 ### Changed
 
 ### Fixed
 
 ### Security
 
-See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-10-15.903...main) for everything awaiting release
+See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-10-17.921...main) for everything awaiting release
+
+---
+
+## [1.22.2](https://github.com/DFE-Digital/complete-conversions-transfers-changes/releases/tag/production-2025-10-17.921) - 2025-10-17
+
+### Added
+- Added `Confirm the headteacher's details` task for both conversion and transfer projects.
+
+### Changed
+
+### Fixed
+- make internal email check case insensitive
+
+### Security
+
+See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-10-16.912...production-2025-10-17.921) for everything in the release
+
+---
+
+## [1.22.1](https://github.com/DFE-Digital/complete-conversions-transfers-changes/releases/tag/production-2025-10-16.912) - 2025-10-16
+
+### Changed
+- removed checkbox with no corresponding database field
+
+### Fixed
+- resolve bad field mappings for master funding agreements on conversion projects
+
+See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-10-15.903...production-2025-10-16.912) for everything in the release
 
 ---